move JsonUtil.StringDoubleMapFormat and JsonUtil.StringStringMapFormat direclty into JsonFormats as implicit objects - fixes #36
